new arch largely done, minutae remain
