+++ .coveragerc
@@ -2,3 +2,4 @@
 branch = True
 omit =
 	*/__init__.py
+	*/helpers.py
