+++ TODO.md
@@ -1,7 +1,8 @@
 - Document all remaining filters and functions with docstrings
 - Collate the scientific discussions to be had, including the TBC QC filter, with responsible scientists (these are for a future release)
-- Complete implementation of unit tests for DVFilter
-- Expose additional args to user from DVFilter
+- implement DVF filter read ratio threshold and expose as user param
 - Run prerelease over many samples
 - update README w.r.t to new filter and changes in filter code reporting
 - bump version w.r.t to new filter, and new archtype - since filter reporting is not backwards compatible, this is probably a bump of a major version
+- report (accidental) bugfix of variants with only one supporting read being missed for ALF testing due to incorrect return in alt_filter_reads() - hence higher reports of ALF in new version
+- look for boundary conditions which are untested, or other untested paths which should be tested
