+++ hairpin2/abstractfilters.py
@@ -1,3 +1,26 @@
+# hairpin2
+#
+# Copyright (C) 2024 Genome Research Ltd.
+#
+# Author: Alex Byrne <ab63@sanger.ac.uk>
+#
+# Permission is hereby granted, free of charge, to any person obtaining a copy
+# of this software and associated documentation files (the "Software"), to deal
+# in the Software without restriction, including without limitation the rights
+# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+# copies of the Software, and to permit persons to whom the Software is
+# furnished to do so, subject to the following conditions:
+#
+# The above copyright notice and this permission notice shall be included in all
+# copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+# SOFTWARE.
 # TODO: give example in docstring?
 """
 A set of Parent Abstract Base Classes that together completely describe the expected implementation of a scientific test
@@ -11,7 +34,7 @@ The payoff for that verbosity is:
 from abc import ABC, abstractmethod
 from pydantic import BaseModel, ConfigDict
 from pydantic.dataclasses import dataclass
-from typing import Generic, TypeVar, Any, override
+from typing import Generic, TypeVar, Any, override, ClassVar
 from collections.abc import Collection, Mapping
 from pysam import AlignedSegment
 from enum import IntEnum 
@@ -23,7 +46,7 @@ from enum import IntEnum
 
 CodeEnum_T = TypeVar("CodeEnum_T", bound=IntEnum, covariant=True)
 class FilterResult(BaseModel, Generic[CodeEnum_T], ABC):
-    # TODO: pydantic now obsoletes the Codes classvar, document
+    # TODO: pydantic now obsoletes the Codes classvar, change docstring
     """
     Parent ABC class defining the implementation that must be followed by subclasses intending to hold results of running a `FilterTester.test()` on a variant
     All filters should use subclasses that inherit from this class to hold their results.
@@ -48,7 +71,7 @@ class FilterResult(BaseModel, Generic[CodeEnum_T], ABC):
         ... # the rest of the class body, e.g. further instance variables to be associated with a particular result
     ```
     """
-    name: str  # for VCF FILTER field
+    Name: ClassVar[str]  # for VCF FILTER field
     flag: bool | None
     code: CodeEnum_T | None
 
@@ -60,7 +83,7 @@ class FilterResult(BaseModel, Generic[CodeEnum_T], ABC):
             raise ValueError('If flag is set a code must be provided')
 
     @abstractmethod
-    def getinfo(self) -> str | None:
+    def getinfo(self) -> str:
         """
         Return basic filter info in a string formatted for use in the VCF INFO field - "<flag>|<code>".
