+++ hairpin2/filters/ALF.py
@@ -1,9 +1,31 @@
+# hairpin2
+#
+# Copyright (C) 2024 Genome Research Ltd.
+#
+# Author: Alex Byrne <ab63@sanger.ac.uk>
+#
+# Permission is hereby granted, free of charge, to any person obtaining a copy
+# of this software and associated documentation files (the "Software"), to deal
+# in the Software without restriction, including without limitation the rights
+# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+# copies of the Software, and to permit persons to whom the Software is
+# furnished to do so, subject to the following conditions:
+#
+# The above copyright notice and this permission notice shall be included in all
+# copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+# SOFTWARE.
 import hairpin2.abstractfilters as haf
-from pydantic import Field
-from pydantic.dataclasses import dataclass
-from typing import override
-from collections.abc import Sequence
 from pysam import AlignedSegment
+from pydantic.dataclasses import dataclass
+from typing import override, ClassVar
+from collections.abc import Sequence
 from enum import IntEnum, auto
 from statistics import median
 # If you're here just to examine the scientific implementation of each filter,
@@ -18,9 +40,9 @@ class ALCodes(IntEnum):
 
 
 class Result(haf.FilterResult[ALCodes]):
+    Name: ClassVar[str] = 'ALF'
     alt: str
     avg_as: float | None
-    name: str = Field(default='ALF', init=False)
 
     @override
     def getinfo(self) -> str:
