+++ test/test_flag_read_broad.py
@@ -21,16 +21,13 @@
 # LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
-from hairpin2.main import qc_read_broad
+from hairpin2.readqc import qc_read_broad
 from hairpin2 import constants as cnst
 import pysam
 from pysam.libcalignedsegment import SAM_FLAGS as s
 import copy
 
 
-# BASIS PATH TESTING (ish)
-# test every node and edge at least once
-# ----
 # perfect read pair:
 r = pysam.AlignedSegment()
 r.query_name = 'read1'
