+++ test/test_ref2querypos.py
@@ -26,9 +26,6 @@ from pytest import raises
 import pysam
 
 
-# BASIS PATH TESTING (ish)
-# test every node and edge at least once
-# ----
 # perfect read pair:
 r = pysam.AlignedSegment()
 r.query_name = 'read1'
